Moved auditableSerializable class to other file
